Update resign.py
Removed zipallign, as not needed (for apks are aligned already) and was messing up the signatures.
Added abilityto pass multiple folders to be resined. This allows same signature to be used if apk also exists in other folder (as in vendor folder).
Added ability to use the apksigner command rather than the Java library. You can do that by setting`useApkSigner`  to True.
Other small changes.